complete the separation
