+++ modules/core/src/main/scala/sangria/execution/ExecutionError.scala
@@ -1,42 +1,9 @@
 package sangria.execution
 
 import sangria.ast.AstLocation
-import sangria.marshalling.ResultMarshaller
 import sangria.ast.SourceMapper
 import sangria.schema.{AbstractType, InterfaceType, ObjectType, UnionType}
 import sangria.validation.{AstNodeLocation, Violation}
-
-trait UserFacingError {
-  def getMessage(): String
-}
-
-trait WithViolations extends UserFacingError {
-  def violations: Vector[Violation]
-}
-
-trait ErrorWithResolver {
-  this: Throwable =>
-
-  def exceptionHandler: ExceptionHandler
-
-  def resolveError(implicit marshaller: ResultMarshaller): marshaller.Node =
-    new ResultResolver(marshaller, exceptionHandler, false)
-      .resolveError(this)
-      .asInstanceOf[marshaller.Node]
-}
-
-class ExecutionError(
-    message: String,
-    val exceptionHandler: ExceptionHandler,
-    val sourceMapper: Option[SourceMapper] = None,
-    val locations: List[AstLocation] = Nil)
-    extends Exception(message)
-    with AstNodeLocation
-    with UserFacingError
-    with ErrorWithResolver {
-  override def simpleErrorMessage = super.getMessage
-  override def getMessage() = super.getMessage + astLocation
-}
 
 abstract class InternalExecutionError(message: String)
     extends Exception(message)
@@ -82,24 +49,6 @@ case object IntrospectionNotAllowedError
     extends Exception(s"Introspection is not allowed.")
     with UserFacingError
 
-trait QueryAnalysisError extends ErrorWithResolver {
-  this: Throwable =>
-}
-
-case class VariableCoercionError(violations: Vector[Violation], eh: ExceptionHandler)
-    extends ExecutionError(
-      s"Error during variable coercion. Violations:\n\n${violations.map(_.errorMessage).mkString("\n\n")}",
-      eh)
-    with WithViolations
-    with QueryAnalysisError
-
-case class AttributeCoercionError(violations: Vector[Violation], eh: ExceptionHandler)
-    extends ExecutionError(
-      s"Error during attribute coercion. Violations:\n\n${violations.map(_.errorMessage).mkString("\n\n")}",
-      eh)
-    with WithViolations
-    with QueryAnalysisError
-
 case class ValidationError(violations: Vector[Violation], eh: ExceptionHandler)
     extends ExecutionError(
       s"Query does not pass validation. Violations:\n\n${violations.map(_.errorMessage).mkString("\n\n")}",
@@ -121,10 +70,6 @@ case class MaterializedSchemaValidationError(
       s"Materialized schema does not pass validation. Violations:\n\n${violations.map(_.errorMessage).mkString("\n\n")}",
       eh)
     with WithViolations
-    with QueryAnalysisError
-
-case class QueryReducingError(cause: Throwable, exceptionHandler: ExceptionHandler)
-    extends Exception(s"Query reducing error: ${cause.getMessage}", cause)
     with QueryAnalysisError
 
 case class OperationSelectionError(
