+++ modules/schema/src/main/scala/sangria/execution/Extension.scala
@@ -1,0 +1,5 @@
+package sangria.execution
+
+import sangria.marshalling.InputUnmarshaller
+
+case class Extension[In](data: In)(implicit val iu: InputUnmarshaller[In])
