+++ modules/schema/src/main/scala/sangria/execution/FieldTag.scala
@@ -2,6 +2,15 @@ package sangria.execution
 
 import sangria.streaming.SubscriptionStream
 
+import scala.language.implicitConversions
+
 trait FieldTag
+
+case class StringTag(name: String)
+
+object FieldTag {
+  implicit def stringTag(s: String): StringTag = StringTag(s)
+  implicit def symbolTag(s: Symbol): StringTag = StringTag(s.name)
+}
 
 case class SubscriptionField[S[_]](stream: SubscriptionStream[S]) extends FieldTag
