+++ modules/schema/src/main/scala/sangria/execution/MiddlewareAttachment.scala
@@ -1,0 +1,3 @@
+package sangria.execution
+
+trait MiddlewareAttachment
