+++ modules/schema/src/main/scala/sangria/execution/ValueCollector.scala
@@ -135,7 +135,7 @@ object ValueCollector {
           val astValue = astArgMap.get(argDef.name).map(_.value)
           val fromInput = argDef.fromInput
 
-          implicit val um = sangria.marshalling.queryAst.queryAstInputUnmarshaller
+          implicit val um = sangria.marshalling.queryAstCore.queryAstInputUnmarshaller
 
           try resolveMapValue(
             argDef.argumentType,
@@ -195,24 +195,4 @@ object ValueCollector {
       }
     }
   }
-}
-
-case class VariableValue(
-    fn: (
-        ResultMarshaller,
-        ResultMarshaller,
-        InputType[_]) => Either[Vector[Violation], Trinary[ResultMarshaller#Node]]) {
-  private val cache =
-    Cache.empty[(Int, Int), Either[Vector[Violation], Trinary[ResultMarshaller#Node]]]
-
-  def resolve(
-      marshaller: ResultMarshaller,
-      firstKindMarshaller: ResultMarshaller,
-      actualType: InputType[_]): Either[Vector[Violation], Trinary[firstKindMarshaller.Node]] =
-    cache
-      .getOrElseUpdate(
-        System.identityHashCode(firstKindMarshaller) -> System.identityHashCode(
-          actualType.namedType),
-        fn(marshaller, firstKindMarshaller, actualType))
-      .asInstanceOf[Either[Vector[Violation], Trinary[firstKindMarshaller.Node]]]
 }
