+++ modules/schema/src/main/scala/sangria/renderer/SchemaRenderer.scala
@@ -3,8 +3,7 @@ package sangria.renderer
 import sangria.ast
 import sangria.execution.ValueCoercionHelper
 import sangria.introspection._
-import sangria.marshalling.{InputUnmarshaller, ToInput}
-import sangria.parser.QueryParser
+import sangria.marshalling.ToInput
 import sangria.schema._
 
 object SchemaRenderer {
@@ -55,21 +54,11 @@ object SchemaRenderer {
       case IntrospectionNamedTypeRef(_, name) => ast.NamedType(name)
     }
 
-  def renderDefault(defaultValue: Option[String]) =
-    defaultValue.flatMap(d => QueryParser.parseInput(d).toOption)
-
   def renderDefault(value: (Any, ToInput[_, _]), tpe: InputType[_]) = {
     val coercionHelper = new ValueCoercionHelper[Any]
 
     DefaultValueRenderer.renderInputValue(value, tpe, coercionHelper)
   }
-
-  def renderArg(arg: IntrospectionInputValue) =
-    ast.InputValueDefinition(
-      arg.name,
-      renderTypeName(arg.tpe),
-      renderDefault(arg.defaultValue),
-      description = renderDescription(arg.description))
 
   def renderArg(arg: Argument[_]) =
     ast.InputValueDefinition(
@@ -92,32 +81,14 @@ object SchemaRenderer {
       case _ => Vector.empty
     }
 
-  def renderArgsI(args: Seq[IntrospectionInputValue]) =
-    args.map(renderArg).toVector
-
   def renderArgs(args: Seq[Argument[_]]) =
     args.map(renderArg).toVector
-
-  def renderFieldsI(fields: Seq[IntrospectionField]) =
-    fields.map(renderField).toVector
 
   def renderFields(fields: Seq[Field[_, _]]) =
     fields.map(renderField).toVector
 
-  def renderInputFieldsI(fields: Seq[IntrospectionInputValue]) =
-    fields.map(renderInputField).toVector
-
   def renderInputFields(fields: Seq[InputField[_]]) =
     fields.map(renderInputField).toVector
-
-  def renderField(field: IntrospectionField) =
-    ast.FieldDefinition(
-      field.name,
-      renderTypeName(field.tpe),
-      renderArgsI(field.args),
-      renderDeprecation(field.isDeprecated, field.deprecationReason),
-      renderDescription(field.description)
-    )
 
   def renderField(field: Field[_, _]) =
     ast.FieldDefinition(
@@ -130,13 +101,6 @@ object SchemaRenderer {
       renderDescription(field.description)
     )
 
-  def renderInputField(field: IntrospectionInputValue) =
-    ast.InputValueDefinition(
-      field.name,
-      renderTypeName(field.tpe),
-      renderDefault(field.defaultValue),
-      description = renderDescription(field.description))
-
   def renderInputField(field: InputField[_]) =
     ast.InputValueDefinition(
       field.name,
@@ -145,13 +109,6 @@ object SchemaRenderer {
       field.astDirectives,
       renderDescription(field.description)
     )
-
-  def renderObject(tpe: IntrospectionObjectType) =
-    ast.ObjectTypeDefinition(
-      tpe.name,
-      renderImplementedInterfaces(tpe),
-      renderFieldsI(tpe.fields),
-      description = renderDescription(tpe.description))
 
   def renderObject(tpe: ObjectType[_, _]) =
     ast.ObjectTypeDefinition(
@@ -200,24 +157,12 @@ object SchemaRenderer {
   def renderScalar(tpe: ScalarType[_]) =
     ast.ScalarTypeDefinition(tpe.name, tpe.astDirectives, renderDescription(tpe.description))
 
-  def renderInputObject(tpe: IntrospectionInputObjectType) =
-    ast.InputObjectTypeDefinition(
-      tpe.name,
-      renderInputFieldsI(tpe.inputFields),
-      description = renderDescription(tpe.description))
-
   def renderInputObject(tpe: InputObjectType[_]) =
     ast.InputObjectTypeDefinition(
       tpe.name,
       renderInputFields(tpe.fields),
       tpe.astDirectives,
       renderDescription(tpe.description))
-
-  def renderInterface(tpe: IntrospectionInterfaceType) =
-    ast.InterfaceTypeDefinition(
-      tpe.name,
-      renderFieldsI(tpe.fields),
-      description = renderDescription(tpe.description))
 
   def renderInterface(tpe: InterfaceType[_, _]) =
     ast.InterfaceTypeDefinition(
@@ -238,25 +183,6 @@ object SchemaRenderer {
       tpe.types.map(t => ast.NamedType(t.name)).toVector,
       tpe.astDirectives,
       renderDescription(tpe.description))
-
-  private def renderSchemaDefinition(schema: IntrospectionSchema): Option[ast.SchemaDefinition] =
-    if (isSchemaOfCommonNames(
-        schema.queryType.name,
-        schema.mutationType.map(_.name),
-        schema.subscriptionType.map(_.name)))
-      None
-    else {
-      val withQuery = Vector(
-        ast.OperationTypeDefinition(ast.OperationType.Query, ast.NamedType(schema.queryType.name)))
-      val withMutation = schema.mutationType.fold(withQuery)(t =>
-        withQuery :+ ast.OperationTypeDefinition(ast.OperationType.Mutation, ast.NamedType(t.name)))
-      val withSubs = schema.subscriptionType.fold(withMutation)(t =>
-        withMutation :+ ast.OperationTypeDefinition(
-          ast.OperationType.Subscription,
-          ast.NamedType(t.name)))
-
-      Some(ast.SchemaDefinition(withSubs, description = renderDescription(schema.description)))
-    }
 
   private def renderSchemaDefinition(schema: Schema[_, _]): Option[ast.SchemaDefinition] =
     if (isSchemaOfCommonNames(
@@ -279,23 +205,12 @@ object SchemaRenderer {
         ast.SchemaDefinition(withSubs, schema.astDirectives, renderDescription(schema.description)))
     }
 
-  private def isSchemaOfCommonNames(
+  private[renderer] def isSchemaOfCommonNames(
       query: String,
       mutation: Option[String],
       subscription: Option[String]) =
     query == "Query" && mutation.fold(true)(_ == "Mutation") && subscription.fold(true)(
       _ == "Subscription")
-
-  def renderType(tpe: IntrospectionType): ast.TypeDefinition =
-    tpe match {
-      case o: IntrospectionObjectType => renderObject(o)
-      case u: IntrospectionUnionType => renderUnion(u)
-      case i: IntrospectionInterfaceType => renderInterface(i)
-      case io: IntrospectionInputObjectType => renderInputObject(io)
-      case s: IntrospectionScalarType => renderScalar(s)
-      case e: IntrospectionEnumType => renderEnum(e)
-      case kind => throw new IllegalArgumentException(s"Unsupported kind: $kind")
-    }
 
   def renderType(tpe: Type with Named): ast.TypeDefinition =
     tpe match {
@@ -318,46 +233,6 @@ object SchemaRenderer {
       renderArgs(dir.arguments),
       dir.locations.toVector.map(renderDirectiveLocation).sortBy(_.name),
       renderDescription(dir.description))
-
-  def renderDirective(dir: IntrospectionDirective) =
-    ast.DirectiveDefinition(
-      dir.name,
-      renderArgsI(dir.args),
-      dir.locations.toVector.map(renderDirectiveLocation).sortBy(_.name),
-      renderDescription(dir.description))
-
-  def schemaAstFromIntrospection(
-      introspectionSchema: IntrospectionSchema,
-      filter: SchemaFilter = SchemaFilter.default): ast.Document = {
-    val schemaDef = if (filter.renderSchema) renderSchemaDefinition(introspectionSchema) else None
-    val types = introspectionSchema.types
-      .filter(t => filter.filterTypes(t.name))
-      .sortBy(_.name)
-      .map(renderType)
-    val directives = introspectionSchema.directives
-      .filter(d => filter.filterDirectives(d.name))
-      .sortBy(_.name)
-      .map(renderDirective)
-
-    ast.Document(schemaDef.toVector ++ types ++ directives)
-  }
-
-  def renderSchema(introspectionSchema: IntrospectionSchema): String =
-    QueryRenderer.renderPretty(
-      schemaAstFromIntrospection(introspectionSchema, SchemaFilter.default))
-
-  def renderSchema[T: InputUnmarshaller](introspectionResult: T): String =
-    QueryRenderer.renderPretty(
-      schemaAstFromIntrospection(
-        IntrospectionParser.parse(introspectionResult).get,
-        SchemaFilter.default))
-
-  def renderSchema(introspectionSchema: IntrospectionSchema, filter: SchemaFilter): String =
-    QueryRenderer.renderPretty(schemaAstFromIntrospection(introspectionSchema, filter))
-
-  def renderSchema[T: InputUnmarshaller](introspectionResult: T, filter: SchemaFilter): String =
-    QueryRenderer.renderPretty(
-      schemaAstFromIntrospection(IntrospectionParser.parse(introspectionResult).get, filter))
 
   def schemaAst(schema: Schema[_, _], filter: SchemaFilter = SchemaFilter.default): ast.Document = {
     val schemaDef = if (filter.renderSchema) renderSchemaDefinition(schema) else None
